chat bot backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncmy==0.2.11
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
flake8==7.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsql==0.91
MarkupSafe==3.0.3
mccabe==0.7.0
mysql-connector-python==9.5.0
packaging==26.0
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.2
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
groq>=0.9.0
python-dotenv>=1.0.0